Extract `option` package in `:model`
