+++ model/src/main/kotlin/io/spine/validation/ValidationPlugin.kt
@@ -31,19 +31,37 @@ import io.spine.protodata.plugin.Policy
 import io.spine.protodata.plugin.View
 import io.spine.protodata.plugin.ViewRepository
 import io.spine.protodata.render.Renderer
-import io.spine.validation.bound.MaxFieldView
-import io.spine.validation.bound.MaxPolicy
-import io.spine.validation.bound.MinFieldView
-import io.spine.validation.bound.MinPolicy
-import io.spine.validation.bound.RangeFieldView
-import io.spine.validation.bound.RangePolicy
-import io.spine.validation.required.IfMissingPolicy
-import io.spine.validation.required.RequireMessageView
-import io.spine.validation.required.RequirePolicy
-import io.spine.validation.required.RequiredFieldView
-import io.spine.validation.required.RequiredIdOptionPolicy
-import io.spine.validation.required.RequiredIdPatternPolicy
-import io.spine.validation.required.RequiredPolicy
+import io.spine.validation.option.bound.MaxFieldView
+import io.spine.validation.option.bound.MaxPolicy
+import io.spine.validation.option.bound.MinFieldView
+import io.spine.validation.option.bound.MinPolicy
+import io.spine.validation.option.bound.RangeFieldView
+import io.spine.validation.option.bound.RangePolicy
+import io.spine.validation.option.ChoiceGroupView
+import io.spine.validation.option.ChoicePolicy
+import io.spine.validation.option.DistinctFieldView
+import io.spine.validation.option.DistinctPolicy
+import io.spine.validation.option.GoesFieldView
+import io.spine.validation.option.GoesPolicy
+import io.spine.validation.option.IfHasDuplicatesPolicy
+import io.spine.validation.option.IfInvalidPolicy
+import io.spine.validation.option.IfSetAgainPolicy
+import io.spine.validation.option.IsRequiredPolicy
+import io.spine.validation.option.PatternFieldView
+import io.spine.validation.option.PatternPolicy
+import io.spine.validation.option.SetOnceFieldView
+import io.spine.validation.option.SetOncePolicy
+import io.spine.validation.option.ValidatePolicy
+import io.spine.validation.option.ValidatedFieldView
+import io.spine.validation.option.WhenFieldView
+import io.spine.validation.option.WhenPolicy
+import io.spine.validation.option.required.IfMissingPolicy
+import io.spine.validation.option.required.RequireMessageView
+import io.spine.validation.option.required.RequirePolicy
+import io.spine.validation.option.required.RequiredFieldView
+import io.spine.validation.option.required.RequiredIdOptionPolicy
+import io.spine.validation.option.required.RequiredIdPatternPolicy
+import io.spine.validation.option.required.RequiredPolicy
 
 /**
  * The basic implementation of ProtoData validation plugin, which builds
