+++ model/src/main/kotlin/io/spine/validation/option/bound/BoundFieldSupport.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.bound
+package io.spine.validation.option.bound
 
 import io.spine.protodata.Compilation
 import io.spine.protodata.ast.Field
