+++ model/src/main/kotlin/io/spine/validation/option/bound/MinOption.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.bound
+package io.spine.validation.option.bound
 
 import io.spine.core.External
 import io.spine.core.Subscribe
@@ -49,7 +49,8 @@ import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validation.MIN
 import io.spine.validation.RANGE
 import io.spine.validation.api.OPTION_NAME
-import io.spine.validation.bound.BoundFieldSupport.checkFieldType
+import io.spine.validation.bound.MinField
+import io.spine.validation.option.bound.BoundFieldSupport.checkFieldType
 import io.spine.validation.defaultMessage
 import io.spine.validation.bound.event.MinFieldDiscovered
 import io.spine.validation.bound.event.minFieldDiscovered
